Data packing from @0xth0mas

+++ contracts/ILifeOnchain.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.17;
  * @title Traits indexes for each life
  */
 struct Traits {
-    uint16 speedIndex;
-    uint16 colorIndex;
-    uint16 modeIndex;
+    uint256 speedIndex;
+    uint256 colorIndex;
+    uint256 modeIndex;
 }
